@@ -77,13 +77,13 @@
     "lodash.keyby": "4.6.0",
     "moment": "2.13.0",
     "numeral": "1.5.3",
-    "nvd3": "1.8.1",
-    "react": "15.3.0",
-    "react-dom": "15.3.0",
+    "nvd3": "1.8.5",
+    "react": "15.6.1",
+    "react-dom": "15.6.1",
     "react-nvd3": "0.5.7",
-    "react-redux": "5.0.5",
-    "react-sizeme": "2.2.0",
-    "react-sparklines": "1.4.2",
+    "react-redux": "5.0.6",
+    "react-sizeme": "2.3.4",
+    "react-sparklines": "1.7.0",
     "redux": "3.7.2",
     "redux-actions": "^2.2.1",
     "redux-observable": "0.14.1",
@@ -97,11 +97,10 @@
     "@types/jest": "20.0.5",
     "@types/karma": "0.13.36",
     "@types/lodash": "3.10.1",
-    "@types/react": "^15.0.39",
-    "@types/react-dom": "^15.5.1",
+    "@types/react": "16.0.2",
+    "@types/react-dom": "15.5.2",
     "@types/react-virtualized": "9.7.3",
-    "@types/sinon": "^2.3.3",
-    "@types/redux-actions": "^1.2.7",
+    "@types/redux-actions": "1.2.7",
     "autoprefixer": "6.3.1",
     "awesome-typescript-loader": "3.2.2",
     "babel-core": "^6.25.0",
@@ -129,10 +128,10 @@
     "node-sass": "^4.5.3",
     "npm-run-all": "2.1.0",
     "openfin-cli": "1.0.2",
-    "react-addons-test-utils": "15.3.0",
-    "react-test-renderer": "^15.6.1",
-    "react-transform-catch-errors": "1.0.1",
-    "react-transform-hmr": "1.0.1",
+    "react-addons-test-utils": "15.6.0",
+    "react-test-renderer": "15.6.1",
+    "react-transform-catch-errors": "1.0.2",
+    "react-transform-hmr": "1.0.4",
     "react-virtualized": "9.9.0",
     "redbox-react": "1.2.0",
     "redux-devtools-extension": "2.13.2",
