@@ -1,7 +1,6 @@
-import {Observable} from 'rxjs/Observable'
 import * as _ from 'lodash'
 import * as keyBy from 'lodash.keyby'
-import {createAction} from 'redux-actions'
+import { createAction } from 'redux-actions';
 
 import {ACTION_TYPES as REGION_ACTION_TYPES} from '../regions/regionsOperations'
 import {ACTION_TYPES as REF_ACTION_TYPES} from '../reference/referenceOperations'
@@ -18,7 +17,7 @@ export const onComponentMount = createAction(REGION_ACTION_TYPES.REGION_ADD, pay
 
 export const blotterServiceEpic = blotterService$ => action$ => {
   return action$.ofType(REF_ACTION_TYPES.REFERENCE_SERVICE)
-    .mergeMapTo(Observable.merge(blotterService$.getTradesStream()))
+    .flatMapTo(blotterService$.getTradesStream())
     .map(fetchBlotter);
 }
 
@@ -29,9 +28,10 @@ export const blotterServiceReducer = (state: any = {}, action) => {
     case ACTION_TYPES.BLOTTER_SERVICE:
       const trades = keyBy(_.values(action.payload.trades), '_tradeId')
       return {
-        trades: _.sortBy({...trades, ...state.trades}, '_topicId').reverse()
+        trades: _.sortBy({ ...trades, ...state.trades }, '_topicId').reverse()
       }
     default:
       return state
   }
 }
+
