@@ -1,10 +1,13 @@
 import { combineReducers } from 'redux'
 
 import { blotterServiceReducer } from './blotter/blotterOperations'
-import { referenceServiceReducer } from './reference/referenceOperations'
-import { pricingServiceReducer } from './pricing/pricingOperations'
-import { analyticsServiceReducer } from './analytics/analyticsOperations'
-import { compositeStatusServiceReducer } from './compositeStatusService/compositeStatusServiceOperations'
+import { referenceServiceReducer } from './reference/referenceOperations';
+import { pricingServiceReducer } from './pricing/pricingOperations';
+import compositeStatusServiceReducer from './compositeStatusService/compositeStatusServiceOperations';
+import connectionStatusReducer from './connectionStatus/connectionStatusOperations';
+import analyticsServiceReducer from './analytics/analyticsOperations';
+import sidebarRegionReducer from '../ui/sidebar/SidebarRegionOperations';
+import footerReducer from '../ui/footer/FooterOperations';
 import { regionsReducer } from "./regions/regionsOperations"
 
 
@@ -14,6 +17,9 @@ const rootReducer = combineReducers({
   pricingService: pricingServiceReducer,
   analyticsService: analyticsServiceReducer,
   compositeStatusService: compositeStatusServiceReducer,
+  connectionStatus: connectionStatusReducer,
+  displayAnalytics: sidebarRegionReducer,
+  displayStatusServices: footerReducer,
   regionsService: regionsReducer
 })
 
