@@ -8,26 +8,27 @@ import { referenceServiceEpic } from './reference/referenceOperations'
 import { pricingServiceEpic } from './pricing/pricingOperations'
 import { analyticsServiceEpic } from './analytics/analyticsOperations';
 import { compositeStatusServiceEpic } from './compositeStatusService/compositeStatusServiceOperations';
+import { connectionStatusEpic } from './connectionStatus/connectionStatusOperations';
 import { openFinEpic } from './openFinEpic'
 
-
-const creatEpicMiddleware = (referenceDataService, blotterService, pricingService, analyticsService, compositeStatusService, openFin) => createEpicMiddleware(
+const epicMiddleware = (referenceDataService, blotterService, pricingService, analyticsService, compositeStatusService, openFin) => createEpicMiddleware(
   combineEpics(
     referenceServiceEpic(referenceDataService),
     blotterServiceEpic(blotterService),
     pricingServiceEpic(pricingService),
     analyticsServiceEpic(analyticsService),
     compositeStatusServiceEpic(compositeStatusService),
+    connectionStatusEpic(compositeStatusService),
     openFinEpic(openFin)
   )
 )
 
 export default function configureStore(referenceDataService, blotterService, pricingService, analyticsService, compositeStatusService, openFin) {
-  const epicMiddleware = creatEpicMiddleware(referenceDataService, blotterService, pricingService, analyticsService, compositeStatusService, openFin)
+  const middleware = epicMiddleware(referenceDataService, blotterService, pricingService, analyticsService, compositeStatusService, openFin)
   return createStore(
     rootReducer,
     composeWithDevTools(
-      applyMiddleware(epicMiddleware)
+      applyMiddleware(middleware)
     )
   )
 }
